BUG:Fixed warnings

+++ VTK/Hybrid/vtkBarChartActor.cxx
@@ -517,7 +517,7 @@ int vtkBarChartActor::PlaceAxes(vtkViewport *viewport, int *vtkNotUsed(size))
       }
     else
       {
-      sprintf(label,"%d",i);
+      sprintf(label,"%d",static_cast<int>(i));
       this->LegendActor->SetEntryString(i,label);
       }
     }
@@ -537,7 +537,7 @@ int vtkBarChartActor::PlaceAxes(vtkViewport *viewport, int *vtkNotUsed(size))
         }
       else
         {
-        sprintf(label,"%d",i);
+        sprintf(label,"%d",static_cast<int>(i));
         this->BarMappers[i]->SetInput(label);
         }
       this->BarMappers[i]->GetTextProperty()->ShallowCopy(this->LabelTextProperty);
